tests/wpt/mozilla/tests/mozilla/mozbrowser/crossorigin_xhr.html

<head>
<title>cross origin xhr() with mozbrowser</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>

async_test(function(t) {
  var xhr = new XMLHttpRequest();
  xhr.open("GET", "http://www2.web-platform.test:8000");
  xhr.send();

  xhr.onerror = this.step_func(() => {
    assert_unreached("Cross origin xhr() should not have failed");
    t.done();
  });

  xhr.onload = this.step_func(() => {
    assert_equals(xhr.status, 200, "Cross origin xhr() is successful");
    t.done();
  });
});

</script>
</body>
